Let support double for real inputs.

+++ 3. Calculations/java/Exercice 7/src/com/bravedroid/Businesslogic/AreaCalculator.java
@@ -1,12 +1,11 @@
 package com.bravedroid.Businesslogic;
 
 public class AreaCalculator {
-
-    public int calculateAreaInFeet(int length, int width) {
+    public double calculateAreaInFeet(double length, double width) {
         return length * width;
     }
 
-    public double calculateAreaInMeters(int areaInFeet, double conversion) {
+    public double calculateAreaInMeters(double areaInFeet, double conversion) {
         return areaInFeet * conversion;
     }
 }
